app/pyproject.toml
[project]
name = "app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.1.1",
    "psycopg2-binary>=2.9.10",
    "pydantic-settings>=2.10.1",
]
